app/requirements-dev.txt
black==22.10.0
cachetools==5.2.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
isort==5.10.1
psycopg2-binary==2.9.5
pytest==7.2.0
requests==2.28.1
SQLAlchemy==1.4.44
SQLAlchemy-Utils==0.38.3